Require a '2nd expansion round' on %{?$python..} macros
When the spec file is parsed, macros are expanded, leaving any
unknown macros intact. This is where %python_expand kicks in,
does its magic to replace '$python' in macros and runs expand
on them again.
However, macros with %{?...} are gone at this point already
if not know at time of the first expansion - which is the case
for the %{?$python_..} contraption.
This can be avoided by explicitly requiring a 2nd pass using
%%{?$python...} instead.
See discussion in bsc#1183616.

Signed-off-by: Egbert Eich <eich@suse.com>