Fix master package README generation 'trick' for rpm <= 4.19.1
rpm version 4.19.1 started to add quotes around the file name
when it expands a %doc statements into a script. This broke
the 'trick' we've used to generate a README file on-the-fly.
Here the saving grace was the introduction of the macro %rpmversion
with the same version of rpm that can be used to identify versions
affected by this.

Signed-off-by: Egbert Eich <eich@suse.com>